Fix Issue #399, clamp minion stats after changing their classes
